cleaned some things up
