+++ genny/model/templates/-package-/-name-.go.tmpl
@@ -4,7 +4,7 @@ import (
 	{{ range $i := .model.Imports -}}
 	"{{$i}}"
 	{{ end -}}
-	{{ if .model.ValidatableAttributes -}}
+	{{ if .model.Validations -}}
 	"github.com/gobuffalo/validate/validators"
 	{{ end -}}
 )
@@ -33,9 +33,9 @@ func ({{.model.Name.Char}} {{.model.Name.Proper.Pluralize}}) String() string {
 // Validate gets run every time you call a "pop.Validate*" (pop.ValidateAndSave, pop.ValidateAndCreate, pop.ValidateAndUpdate) method.
 // This method is not required and may be deleted.
 func ({{.model.Name.Char}} *{{.model.Name.Proper}}) Validate(tx *pop.Connection) (*validate.Errors, error) {
-	{{ if .model.ValidatableAttributes -}}
+	{{ if .model.Validations -}}
 	return validate.Validate(
-		{{ range $a := .model.ValidatableAttributes -}}
+		{{ range $a := .model.Validations -}}
 		&validators.{{capitalize $a.GoType}}IsPresent{Field: {{$.model.Name.Char}}.{{$a.Name.Pascalize}}, Name: "{{$a.Name.Pascalize}}"},
 		{{end -}}
 	), nil
